+++ docs/CHANGELOG.md
@@ -3,9 +3,10 @@
 # Next:
 
 * Package `backends`:
-  * Added `simplego`, a portable, simple albeit very slow backend.
+  * Added `simplego`, a portable, simple albeit slow backend.
+    * Implemented ~50 most common ops, see `backends/simplego/capabilities`, and most common numeric types (including BFloat16).
   * Added sub-package `notimplemented`: helper to implement new backends.
-  * Added sub-package `shapeinference`: helper to impelment new backends.
+  * Added sub-package `shapeinference`: helper to implement new backends.
 * Package `checkpoints`
   * Added `Config.FromEmbed` that allows loading a checkpoint from an embedded variable.
 * Package `graph`: